Convert all watch() to useWatch():
"The only difference between useWatch and watch is at the root (useForm) level or the custom hook level update."

(From: https://react-hook-form.com/docs/usewatch)

This was a very important change that instantly stopped unnecessary re-renders on the whole Property page.

From now on, every component should be using useWatch instead of watch() but it is not a big problem for the rest of the pages (only for property).